update download manager

+++ appcore/src/main/java/mobi/cangol/mobile/utils/StringUtils.java
@@ -95,7 +95,7 @@ public class StringUtils {
 	 * @param value
 	 * @return
 	 */
-	public static String formatSpeed(int value) {
+	public static String formatSpeed(long value) {
 		return formatSize(value) + "/s";
 	}
 	/**
